backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]